CPR-61
. Added find defendants by pnc number
. Added tests
. Refactored repository name.
